requirements.txt
grpcio==1.64.1
grpcio-tools==1.64.1
protobuf==5.27.2
six==1.15.0
termcolor==1.1.0